% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMRmix.R
\name{SMRmix}
\alias{SMRmix}
\title{Mixed Effect Similarity Matrix Regression Model (SMRmix)}
\usage{
SMRmix(
  formula,
  data = NULL,
  Kernels,
  times = 2000,
  var.type = "different",
  outcome = "continuous"
)
}
\arguments{
\item{formula}{A two-sided linear formula describing both the fixed-effects 
and random-effects part of the model.}

\item{data}{An optional data frame containing the variables named in \code{formula}.}

\item{Kernels}{A list that either consists of multiple lists of kernel matrices, or 
a single list of kernel matrices. Denote Kernels\eqn{=\{K_1,\cdots,K_P\}}, where 
\eqn{P} is the number of kernel metrics under consideration. Each \eqn{K_p} (\eqn{p\in \{1,\cdots,P\}}) 
is itself a list of \eqn{K} kernel matrices, where \eqn{K} is the number of 
heterogeneous studies. Each study \eqn{k} (\eqn{k\in \{1,\cdots,K\}}) contributes a kernel
matrix of dimension \eqn{n_k} by \eqn{n_k} to each \eqn{K_p}.}

\item{times}{Number of perturbation replicates. Defaults to \code{2000}.}

\item{var.type}{A character indicating the model assumption for the variance of
study-specific random effect for continuous outcomes. The default option (\code{'different'}) 
assumes the variance \eqn{\delta_k^2} vary across different studies. The option 
\code{var.type='same'} assumes a constant variance across all studies.}

\item{outcome}{A character indicating whether the outcome of interest is \code{continuous}
or \code{binary}. Default to \code{continuous}.}

\item{otu.tab}{Numeric matrix of template microbiome count dataset. Rows are
samples, columns are taxa.}

\item{n.break.ties}{Number of replicates to break ties when ranking relative
abundances. Defaults to \code{100}.}

\item{mode}{A character indicating the modeling approach for relative abundances.
If \code{'parametric'}, a parametric model involving fitting a generalized
gamma distribution is used. If \code{'nonparametric'}, the nonparametric
approach involving quantile matching is applied. Note that a parametric
model is required if library sizes or characteristics of taxa will be modified.
Defaults to \code{'nonparametric'}.}
}
\value{
Returns a list that has components:

Returns a list that has components:
\item{p_values}{Individual p-values for each candidate kernel metric.}
\item{omnibus_p}{Omnibus p-value considering multiple kernel metrics.}
}
\description{
Integrates heterogeneous microbiome datasets and tests the association between 
microbial composition and an outcome of interest.
}
